// frontend/frontend/app/src/app/model/employeeModel.ts
export class Employee{
    username: string;
    password: string;
    name: string;
    surname: string;
    adress: string;
    mobilePhone: string;
    website:string;
    personalData: string;
    profession: string;
    roomNumber: string;
    status: string;
    picture: string;
    numOfLogin: number;
    employeeFlag: number;

}